Refactor types to make the adapter author experience cleaner.
This doesn't change any functionality, it merely moves things
around.

pkg/aspect, pkg/aspect/listChecker, pkg/aspect/denyChecker
and pkg/registry now only contain interface types relevant
to adapter authors.

pkg/aspectsupport, pkg/aspectsupport/listChecker,
pkg/aspectsupport/denyChecker is strictly for
aspect managers.
