+++ adapter/ipListChecker/BUILD
@@ -12,10 +12,9 @@ go_library(
     deps = [
         "@com_github_golang_glog//:go_default_library",
         "@in_gopkg_yaml_v2//:go_default_library",
-        "//pkg/adapter:go_default_library",
         "//pkg/aspect:go_default_library",
-        "//pkg/aspectsupport:go_default_library",
         "//pkg/aspect/listChecker:go_default_library",
+        "//pkg/registry:go_default_library",
         "@com_github_golang_protobuf//proto:go_default_library",
     ],
 )
