+++ pkg/aspect/BUILD
@@ -2,23 +2,15 @@ package(default_visibility=["//visibility:public"])
 
 licenses(["notice"])
 
-load("@io_bazel_rules_go//go:def.bzl", "go_prefix", "go_library")
+load("@io_bazel_rules_go//go:def.bzl", "go_prefix", "go_library", "go_test")
 
 go_prefix("istio.io/mixer/pkg/aspect")
-
-ISTIO_DEPS = [
-    "//:mixer/api/v1",
-    "//:istio/config/v1",
-    "//:istio/config/v1/aspect/denyChecker",
-    "//:istio/config/v1/aspect/listChecker",
-]
 
 DEPS = [
     "@com_github_golang_protobuf//proto:go_default_library",
     "@com_github_google_go_genproto//googleapis/rpc/code:go_default_library",
-    "//pkg/adapter:go_default_library",
-    "//pkg/attribute:go_default_library",
-    "//pkg/expr:go_default_library"] + ["@com_github_istio_api" + d for d in ISTIO_DEPS]
+    "@com_github_hashicorp_go_multierror//:go_default_library",
+]
 
 go_library(
     name="go_default_library",
@@ -27,4 +19,11 @@ go_library(
         exclude=["test_*.go"],
     ),
     deps=DEPS
+)
+
+go_test(
+    name = "config_error_test",
+    size = "small",
+    srcs = ["configError_test.go"],
+    library = ":go_default_library",
 )
